it has been changed

+++ src/main/java/web/dao/UserDao.java
@@ -6,9 +6,9 @@ import java.util.List;
 public interface UserDao {
 
     List<User> listUsers();
-    
-    User getUser(Long id);
-    void editUser(Long id, User user);
-    void deleteUser(Long id);
+
+    User getUser(int id);
+    void editUser(int id, User user);
+    void deleteUser(int id);
     void addUser(User user);
 }
