+++ src/main/java/web/service/UserService.java
@@ -6,7 +6,7 @@ import java.util.List;
 public interface UserService {
     void addUser(User user);
     List<User> listUsers();
-    User getUser(Long id);
-    void editUser(Long id, User user);
-    void deleteUser(Long id);
+    User getUser(int id);
+    void editUser(int id, User user);
+    void deleteUser(int id);
 }
